chore: Refactored filenaming
