+++ files.go
@@ -7,6 +7,37 @@ import (
 	"net/http"
 )
 
+func Delete(id string) error {
+	jwt, err := findToken()
+	if err != nil {
+		return err
+	}
+	url := fmt.Sprintf("https://api.pinata.cloud/v3/files/%s", id)
+
+	req, err := http.NewRequest("DELETE", url, nil)
+	if err != nil {
+		return errors.Join(err, errors.New("failed to create the request"))
+	}
+	req.Header.Set("Authorization", "Bearer "+string(jwt))
+	req.Header.Set("content-type", "application/json")
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return errors.Join(err, errors.New("failed to send the request"))
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return fmt.Errorf("server Returned an error %d, check CID", resp.StatusCode)
+	}
+
+	fmt.Println("File Deleted")
+
+	return nil
+
+}
+
 func ListFiles(amount string, cid string, name string, status string, offset string) (ListResponse, error) {
 	jwt, err := findToken()
 	if err != nil {
